requirements.txt
streamlit
pandas
# psycopg2
psycopg2-binary
python-dotenv
seaborn
matplotlib
ipykernel
python-dotenv
numpy

#Foundation Models API
anthropic
openai

#torch
torch==2.2.2

#LangChain
langchain
langchain-core
langchain-community
langchain-openai==0.1.7
langchain-anthropic
langchain-chroma==0.1.0
langchain-experimental
# langserve[all]
chromadb ==0.4.24
pgvector
pypdf==4.2.0
PyPDF2

# Llama-index
llama-index
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-vector-stores-postgres

#Hugging Face
huggingface-hub==0.22.2
transformers==4.39.3
sentence-transformers==2.6.1